requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.3.0
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
flask-bcrypt==1.0.1
flask-limiter==3.5.0
requests==2.31.0
